Remove stale fb303 counters
Summary:
When counters are renamed (port, trunk or queues), the current FBOSS logic
creates a new monotonic counter with new name and removes (last reference to)
monotonic counter for the old name.

However, the old counter continues to unncessarily get exported. This patch
fixes it by deleting the old counters.

Differential Revision: D14942193

fbshipit-source-id: 34cf7dd5290ae87a2c8002775e9d0940615c47fb