revisionstore: partial chains can be returned from get_delta_chain
Summary:
In the case where a delta chain is split between several logs, it's possible
that part of it may be removed due to some logs being removed. Instead of
treating this as an error, we can simply return the partial chain, the union
content store will simply continue the delta chain on the next store.

Reviewed By: quark-zju

Differential Revision: D14899943

fbshipit-source-id: 7369ee191dc4b35873344cd13c295c72472e0712