mononoke: Allow the hook tailer to exclude specific changesets
Summary:
Add a new option that allows a comma separated list of changesets
that should be excluded by the hook tailer.

Also add an option that allows the hook_tailer to parse a file containing
changesets that it should ignore.

If both arguments are passed, the collection of excluded changesets are
merged.

This will be used by the conveyor node to provide better signal by ignoring known
hook failures caused by hg hook bugs.

Reviewed By: aslpavel

Differential Revision: D14894072

fbshipit-source-id: b488bb0bb5e05d7313005e04c55568dd4be5b770