tests: set CACHEDIR to $TESTTMP/hgcache instead of $PWD/hgcache
Summary:
On Windows, all the tests that are expecting to find some files in $CACHEDIR
would fail due to the directory not existing. Interestingly enough, printing
$CACHEDIR would print a reasonable path, which is the same as $TESTTMP.
Trying to understand this better, I passed --keep-tmpdir to run-tests and
realized that the "real" $TESTTMP was somewhere in my home directory, while
the real $CACHEDIR was in fact C:\tmp.

I haven't fully understood why, but it looks like $PWD is expanded in C:\tmp,
while $TESTTMP is expanded into something else.

Reviewed By: quark-zju

Differential Revision: D15041274

fbshipit-source-id: 0d167183d74df5f6ab84360c5699e96808fceb9b