revisionstore: don't support delta chains in the indexedlog datastore
Summary:
Delta chains are only created during repack, as every download operation
fetches the full content of the file. Even if we wanted to support them,
interrupted chains adds undesirable complexity as it can lead to chain loops if
we're not careful. Let's just not support delta chains for now to avoid this.

Reviewed By: kulshrax

Differential Revision: D15056861

fbshipit-source-id: 4b0474ce134e946952a70f363190faf50850abe0