spanset: introduce SpanSet
Summary:
The spanset is a set of integer spans. It will be used by some DAG related
operations. It'll be used as a subset of mercurial/smartset.py.

Note: smartset.py also has a Python `spanset` structure. That is different
from this Rust spanset in these ways:

- The Rust set does not preserve ordering.
- The Rust set can have multiple spans, instead of just one.
- The Rust set is less abstract (for now). Its set operations (union, etc.)
  only work on the same type.

This diff adds some initial functions for it.

Reviewed By: sfilipco

Differential Revision: D15004985

fbshipit-source-id: c2e5e2a80e2e4681c2f443e0d8a83dc97f7be371