idmap: implement write operations
Summary:
Complete the IdMap interface so it's usable.

There are 2 possible use patterns:
- On-disk IdMap + In-memory additions. Practically, the server provides an
  on-disk map, and the client might assign missing commits on demand. The
  client still needs to update the IdMap during pull.
- Everything is on-disk. There are no in-memory additions. This is more complex
  because the local commits might become part of the server commits in the
  future, and it might require Ids for those commits to be re-assigned.

I haven't decided which way to go exactly. So let's keep the interface flexible
for both.

That said, I do want to reduce the chance of causing filesystem race conditions
for filesystem writes. In this case, both reads and writes should hold a lock.
So a dedicated type is used to encourage the pattern of:

  - get the dedicated type (and hold the filesystem lock)
  - read, write, sync

Write related methods are not moved to the dedicated type, to cover the
in-memory addition use-case.

Reviewed By: sfilipco

Differential Revision: D15008517

fbshipit-source-id: 5d117ed7f2947aed6ed524a3b5199c071908c4ae