hgsubversion: fix subvertpy bindings to allow replacing directories
Summary:
The existing bindings had a bug where if you replace a directory with a
file or symlink in a single commit, it would tell subversion to add the file
before it told it to delete the directory. This resulted in an error from
subversion and meant we couldn't actually push these commits.

Reviewed By: quark-zju

Differential Revision: D15066581

fbshipit-source-id: 1a1db0fa88cd3456fde45a4aec5d4df01837474d