Introduce a test tool to write stub log entries
Summary:
Stanislau and Kostia mentioned that adding a few more commands for managing the HG sync job in the Mononoke Admin tool would be useful — in particular those commands should interact with the list of log entries and the last sync.

Notably, they were:

- Add support for skipping ahead any blobimport log entries.
- Add a verify command to check that all entries looking forward are consistent — i.e. they're either all blobimport or all non-blobimport.

However, as it stands, it's hard to test those in integration tests, because one can't easily generate stub data to exercise on the admin tool.

Notably, inserting data through the `sqlite3` database command line results in inconsistent values for the `reason` column that don't compare equal to those we query from the Rust side of things.

We could work around this by updating the Rust code to e.g. use LIKE instead of equality checks, but then we're updating the production code exclusively for the purpose of serving tests.

So, instead, this patch proposes introduces a lightweight command line  tool to use in tests that basically just inserts data into the bookmarks database, but from the Rust side of things.

Reviewed By: StanislavGlebik

Differential Revision: D15081758

fbshipit-source-id: 574739d2e19e6ce59d729f1aee73c368612fb92b