mononoke: hg sync job: Split out bundle processing in fetch + sync
Summary:
This separates out the bundle processing into two separate steps: preparing (fetching the data) and syncing (sending it to Mercurial).

The underlying goal here is to extract the 1st part into a buffered stream. Doing so will require a little further rework because we need to attach error handling to the stream (in loop mode) or future (in sync-one mode), which requires further refactoring. I'll do so in a separate diff.

Reviewed By: ikostia

Differential Revision: D15121738

fbshipit-source-id: e88d6d730d413e9c1939eefa72ec18799c8b13be