Add hostname endpoint
Summary:
Given that users typically connect to the API server via a VIP, it is
often unclear which actual host we're connected to. As such, API server
responses should contain the actual hostname of the server.

As a first pass at adding this, let's add a simple endpoint that just returns
the hostname, which can be used to see which hosts the VIP is sending requests
to.

Reviewed By: xavierd

Differential Revision: D15168952

fbshipit-source-id: d6fee5716df975fa1588bbf069ad6e970a779b1e