scratch: specify windows config path
Summary:
Previously we would only look for `/etc/scratch.toml`,
which doesn't exist on windows and would resolve ambiguously
depending on the current working directory drive letter.

This diff sets up scratch to look for a site configuration
in a similar location to where we deploy the watchman
configuration on windows systems.

Reviewed By: simpkins

Differential Revision: D15208158

fbshipit-source-id: 3b7e26f14f5b35c6fe7e196eba4e482b26bdeb01