revisionstore: repack packfiles to get under 50 packfiles.
Summary:
When remotefilelog.fetchpacks is enabled, it's possible that 100 packfiles of
100MB each are present. In this case, every new packfiles that
hg_memcache_client would write will force an incremental repack, which will
only reduce the number of packfiles by a small number.

Let's have a simple heuristic that tries to bring the number of packfiles to be
lower than 50.

Reviewed By: DurhamG

Differential Revision: D15203771

fbshipit-source-id: 18c39487d5ac087d4879004993c1c1add087249c