chg: do not start or use a chg server if the process was running with low priority
Summary:
If a chg server was started by a low priority process, it will inherit the low
priority and can be painfully slow.

On the other hand, if the chg client was low priority, the user or the automation
wants the process to be low priority intentionally. Therefore connecting to an
existing chg server might bypass intention.

Therefore, detect the low priority case, and refuse to start or connect to a
chg server.

Previous attempts to workaround slow chg server were removed as they are
probably no longer necessary.

Reviewed By: singhsrb

Differential Revision: D15223518

fbshipit-source-id: 718e807820d481eac204f2293f949c3b315f923d