mononoke: hg-sync-job: allow for graceful exit
Summary:
This adds support in the hg-sync-job for graceful exits by adding an `--exit-file` parameter. When the file referenced by this parameter exists, the sync job will stop processing new entries.

Note that the sync job will actually exit only once it evaluates the exit condition, i.e. the program will only exit once we get the next entry and decide to stop processing work.

Another way to look at this is that creating the flag file guarantees the sync job will not start any new syncs from that point on, but it doesn't guarantee a timely exit.

That's probably what we want here, but if we also want to guarantee a timely exit, then we can add an additional check for the file existence in `loop_over_log_entries` next to the `loop_forever` check.

 ---

The idea here is to use `touch $THE_FILE` as our exit command in Tupperware to allow for gracefully exiting the process.

Reviewed By: ikostia

Differential Revision: D15263468

fbshipit-source-id: 0a5b04e662e2a4042de9e0c5207f1b1be46d1807